+++ client/src/components/layout/landing-page/AboutUs.jsx
