+++ client/src/components/layout/landing-page/AuthButtons.jsx
