+++ client/src/components/layout/landing-page/Features.jsx
@@ -1,10 +1,10 @@
-import "../../../../assets/css/features.css";
+import "../../../assets/css/features.css";
 import { Box, Grid, Paper, Button, useMediaQuery } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
 import Divider from "../../utils/Divider";
 import React, { useState } from "react";
 import ReactCardFlip from "react-card-flip";
-import { featuresData } from "./data/landing-page";
+import { featuresData } from "../../../components/layout/landing-page/data/landing-page";
 
 // icons
