+++ client/src/components/layout/landing-page/data/landing-page.js
