+++ client/src/pages/auth/Login.jsx
@@ -1,5 +1,5 @@
 import React, { Fragment } from "react";
-import LandingLayout from "../../old-ui-src/components/layout/LandingLayout";
+import LandingLayout from "../../components/layout/LandingLayout";
 import { useLocation } from "react-router-dom";
 
 const LoginPage = () => {
