+++ client/src/pages/auth/Register.jsx
@@ -1,6 +1,6 @@
 import React, { Fragment } from "react";
 import { useLocation } from "react-router-dom";
-import LandingLayout from "../../old-ui-src/components/layout/LandingLayout";
+import LandingLayout from "../../components/layout/LandingLayout";
 
 const RegisterPage = () => {
   const location = useLocation()
